Made changes based on PR feedback
