+++ src/HTTP/DataParser/IDataParser.cs
@@ -1,0 +1,9 @@
+using Chorizo.HTTP.Exchange;
+
+namespace Chorizo.HTTP.DataParser
+{
+    public interface IDataParser
+    {
+        Request Parse(byte[] startLineAndHeadersBytes);
+    }
+}
