+++ src/HTTP/ReqProcessor/IRequestProcessor.cs
@@ -1,0 +1,9 @@
+using Chorizo.HTTP.Exchange;
+
+namespace Chorizo.HTTP.ReqProcessor
+{
+    public interface IRequestProcessor
+    {
+        Response Retrieve(Request req);
+    }
+}
