+++ src/HTTP/SocketReader/ISocketReader.cs
@@ -1,0 +1,9 @@
+using Chorizo.Sockets.CzoSocket;
+
+namespace Chorizo.HTTP.SocketReader
+{
+    public interface ISocketReader
+    {
+        byte[] ReadSocket(IChorizoSocket mockSocket);
+    }
+}
