+++ src/IProtocolConnectionHandler.cs
@@ -1,0 +1,9 @@
+using Chorizo.Sockets.CzoSocket;
+
+namespace Chorizo
+{
+    public interface IProtocolConnectionHandler
+    {
+        void HandleRequest(IChorizoSocket chorizoSocket);
+    }
+}
