--- src/ProtocolHandler/DataParser/IDataParser.cs
@@ -1,7 +1,0 @@
-namespace Chorizo.ProtocolHandler.DataParser
-{
-    public interface IDataParser
-    {
-        Request Parse(byte[] startLineAndHeadersBytes);
-    }
-}
