--- src/ProtocolHandler/IChorizoProtocolConnectionHandler.cs
@@ -1,9 +1,0 @@
-using Chorizo.Sockets.CzoSocket;
-
-namespace Chorizo.ProtocolHandler
-{
-    public interface IChorizoProtocolConnectionHandler
-    {   
-        void HandleRequest(IChorizoSocket chorizoSocket);
-    }
-}
