--- src/ProtocolHandler/ResponseRetriever/IResponseRetriever.cs
@@ -1,9 +1,0 @@
-using Chorizo.ProtocolHandler.DataParser;
-
-namespace Chorizo.ProtocolHandler.ResponseRetriever
-{
-    public interface IResponseRetriever
-    {
-        Response Retrieve(Request req);
-    }
-}
