--- src/ProtocolHandler/SocketReader/IHTTPSocketReader.cs
@@ -1,9 +1,0 @@
-using Chorizo.Sockets.CzoSocket;
-
-namespace Chorizo.ProtocolHandler.SocketReader
-{
-    public interface IHTTPSocketReader
-    {
-        byte[] ReadSocket(IChorizoSocket mockSocket);
-    }
-}
