+++ tests/HTTP/ReqProcessor/RequestProcessorTest.cs
@@ -1,0 +1,58 @@
+using System;
+using Chorizo.HTTP.Exchange;
+using Chorizo.HTTP.ReqProcessor;
+using Chorizo.Logger;
+using Moq;
+using Xunit;
+
+namespace Chorizo.Tests.HTTP.ReqProcessor
+{
+    public class RequestProcessorTest
+    {
+        private readonly Mock<IDateTimeProvider> _mockDateTime;
+        public RequestProcessorTest()
+        {
+            _mockDateTime = new Mock<IDateTimeProvider>();
+            var testTime = new DateTime(1997, 12, 02, 15, 10, 00, DateTimeKind.Utc);
+            _mockDateTime.Setup(dt => dt.Now()).Returns(testTime);
+        }
+
+        [Fact]
+        public void RetrieveTakesARequestForForwardSlashSimpleGetAndReturnsAResponseOfTwoHundredOK()
+        {
+            var testRequest = new Request(
+                "GET",
+                "/simple_get",
+                "HTTP/1.1"
+            );
+
+            var testResponse = new Response("HTTP/1.1", 200, "OK")
+                .AddHeader("Date", "Tue, 02 Dec 1997 15:10:00 GMT");
+
+            var testResponseRetriever = new RequestProcessor(_mockDateTime.Object);
+
+            var result = testResponseRetriever.Retrieve(testRequest);
+
+            Assert.True(testResponse.Equals(result));
+        }
+
+        [Fact]
+        public void RetrieveTakesARequestForAPathOtherThanSimpleGetAndThenReturnsAResponseOfFourOFour()
+        {
+            var testRequest = new Request(
+                "GET",
+                "/",
+                "HTTP/1.1"
+            );
+
+            var testResponse = new Response("HTTP/1.1", 404, "Not Found")
+                .AddHeader("Date", "Tue, 02 Dec 1997 15:10:00 GMT");
+
+            var testResponseRetriever = new RequestProcessor(_mockDateTime.Object);
+
+            var result = testResponseRetriever.Retrieve(testRequest);
+
+            Assert.True(testResponse.Equals(result));
+        }
+    }
+}
